packages/core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "priorstudio-core"
version = "0.4.0"
description = "Core abstractions for PriorStudio — Prior, Model, Eval, Run, and the block registry."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "PriorStudio contributors" }]
dependencies = [
  "pydantic>=2.6",
  "pyyaml>=6.0",
  "numpy>=1.26",
]

[project.optional-dependencies]
torch = ["torch>=2.2"]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.4"]

[tool.hatch.build.targets.wheel]
packages = ["priorstudio_core"]
